requirements.txt
langchain-groq>=0.1.0
langchain-google-genai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
